Split native motion projection helpers
